requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
flask>=3.0.0

# Docker SDK for Python
docker>=7.0.0

# Git operations
gitpython>=3.1.40
PyGithub>=2.1.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0

